Vim is not starting up fast but error free.
UltiSnips ftplugin was causing errors in addition
to a ton of missing symlinks that seem to be
largely remedied. As nvim has proven to be
inconsistent enough, I'm making a gradual step
back to vim for the time being. However, that
means catching back up, and that'll take a bit