Add more docs, remove clutter, learn plugins
I'll be adding a [README at](.config/nvim/README.rst).
The idea is to house an explanation of the layout of a typical
Vim directory tree.

As certain plugin configurations continue to grow, another README
may pop up in (.config/nvim/after); however, my notes on Vim
have grown to the point that it may be worth it to either
concatenate them into a wiki or make a ./docs/ folder.

To that end, I added a screenshot of my setup!

The man.vim syntax file is more than likely going to be rewritten
in it's entirety as an amalgation of what nvim and vim provide.